Added stroke weight/color class and showpage class. Fill will take more work
Added some quick new classes for doing line weights and color as well
showpage.

Fill is trickier, as it requires a modification of every print function
to be able to add a gsave -> fill -> grestore after the points are drawn
but before stroke is called, as both stroke and fill destroy the
existing data once called in postscript. Ill make sure to retain current
functionality as well so the tests aren't broken.